proper abstraction between dropper and formdropper
* added DropContext for dropper-specific contexts
* split out the FormDropper and FormDropContext
