+++ src/main/kotlin/us/timinc/mc/cobblemon/droploottables/api/droppers/DropContext.kt
@@ -1,0 +1,3 @@
+package us.timinc.mc.cobblemon.droploottables.api.droppers
+
+interface DropContext
