+++ src/main/kotlin/us/timinc/mc/cobblemon/droploottables/api/droppers/FormDropContext.kt
@@ -1,0 +1,7 @@
+package us.timinc.mc.cobblemon.droploottables.api.droppers
+
+import com.cobblemon.mod.common.pokemon.FormData
+
+data class FormDropContext(
+    val formData: FormData
+) : DropContext
